[core] BINCV_HOST_DEVICE: one implementation compiled for two targets
The backend forks KERNELS because a device traversal has nothing in common
with a row loop. A border-index helper is not a traversal, and forking those
too was costing the project hand-written twins that must stay bit-exact
forever. reduce.cu had already drifted: it opened by claiming the clip geometry
was "the HOST's own impl::clipRegion -- included, not copied -- so the two
backends cannot drift", while restating it forty lines further down. That
sentence is now true.

BINCV_HOST_DEVICE expands to __host__ __device__ under nvcc and to NOTHING
otherwise. It lives in core/error.hpp because that is the one header every
binCV header already includes, so visibility is structural rather than an
include somebody has to remember.

Seventeen sites, each read and confirmed scalar and traversal-free: srcMax,
quantScale, maj3, thresholdGE, thresholdCutoff, minEigenValue, reflect101Edge,
borderIndex, regionFromExtent, clipColumns, clipRegion, bitsPerWord,
lowBitsMask, assertFailed, and Rect's two constructors and empty(). None takes
a view or a BinMat, none allocates, none loops over pixels, rows or words of an
image. thresholdGE is the furthest from "scalar" -- it dereferences a pointer
in a loop -- but the bound is a plane count and the pointer is a caller-held
word array, which is the argument its docstring now makes.

thresholdCutoff is hoisted out of the #ifdef BINCV_WITH_OPENCV block. The gate
that verifies device kernels configures -DBINCV_USE_OPENCV=OFF, so the cutoff
reduction was unreachable from it and a device threshold could only ever have
been proven equal to a test-local copy -- the two-definitions failure this
backend exists to prevent.

The library stays header-only and dependency-free, proven rather than assumed:
all 43 headers compile as one translation unit under the full warning set in an
env -i shell with nvcc off PATH, and preprocessing yields zero occurrences of
__host__, __device__ or __CUDA. A static_assert enforces that the non-nvcc arm
expands to nothing, and it was watched failing against a doctored definition.

Costs the embedded target nothing: the Cortex-M7 image is byte-identical to
pristine HEAD at 24248 / 84 / 291504.

cuda::rowWords and cuda::rowTailMask remain hand-written device copies.
Folding them into the host helpers means a constexpr-ness change on either the
device forms or two Tier 1 host helpers, and no measurement settles which;
ARCHITECTURE now says so plainly and the CUDA suite asserts them equal across
widths.

+++ include/bincv/ops/descriptor.hpp
@@ -235,7 +235,10 @@ inline unsigned briefAngleBin(float angleRadians) {
     constexpr float kBinsPerRadian = static_cast<float>(kBriefAngleBins) / kTwoPi;
     const float t = angleRadians * kBinsPerRadian + static_cast<float>(kBriefAngleBins);
     const unsigned r = static_cast<unsigned>(t + 0.5f);
-    return r % kBriefAngleBins;
+    // `kBriefAngleBins` is a size_t, so the remainder is one too; the cast says
+    // the narrowing is intended, and it is exact -- a remainder modulo 30 fits
+    // an unsigned at every width this library compiles at.
+    return static_cast<unsigned>(r % kBriefAngleBins);
 }
 
 /// @brief `Bits` comparisons at each of the 30 rotations: ~30 KB at 256 bits,
